Fix CI failing

+++ src/singletermfode/PECE.jl
@@ -316,4 +316,4 @@ function A(j::Int, n::Int, α)
     end
 end
 
-B(j::Int, n::Int, α) = ((n + 1 - j)^α - (n - j)^α) # Moved the h^α/α to the end of predictor: return leftsum + h^α/α*predict
+B(j, n, α) = ((n + 1 - j)^α - (n - j)^α) # Moved the h^α/α to the end of predictor: return leftsum + h^α/α*predict
